+++ docs/src/changelog.md
@@ -15,7 +15,7 @@ arguments.
 * Add [`subset_amsr`](@ref).
 * Add [`toc`](@ref) to provide a table of contents for an [`OA`](@ref) object.
 * Add support for coastline data with type [`Coastline`](@ref) plus functions [`coastline`](@ref), [`plot_coastline`](@ref) and [`plot_coastline!`](@ref).
-* Add support for topography data with type [`Topography`](@ref) plus functions [`get_topography_file`](@ref)  and [`read_topography`](@ref).
+* Add support for topography data with type [`Topography`](@ref) plus functions [`get_topography_file`](@ref), [`read_topography`](@ref), and [`plot_topography`](@ref).
 * Add [`pretty`](@ref), for use in contouring (especially for [`plot_TS`](@ref)).
 * Add [`depth_from_pressure`](@ref), [`z_from_pressure`](@ref), [`pressure_from_depth`](@ref),
   and [`pressure_from_z`](@ref).